% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_gst_brow_2_query.R
\name{browse_msig}
\alias{browse_msig}
\title{Retrieve Gene set Names from MSigDB database
Retrieve gene set names from MSigDB database by the gene set name and collection.
the search filed is gene name.}
\usage{
browse_msig(geneSetName = "", collection = "")
}
\arguments{
\item{geneSetName}{one keyword for gene set name, default is empty}

\item{collection}{one collection, default is empty}
}
\value{
gene set names
}
\description{
Retrieve Gene set Names from MSigDB database
Retrieve gene set names from MSigDB database by the gene set name and collection.
the search filed is gene name.
}
\examples{
\donttest{
# missing genSetName and collection to get all gene set names
x <- browse_msig()
# search for gene names include immune
x <- browse_msig('immune')

x |>
    msig_view('cells','response','to','m')

# search for gene names include immune in c8
browse_msig('immune','c8')

# gene names in c8
browse_msig('immune','c8')
}


}
