% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_gst_brow_1_show_collection.R
\name{browse_show_collection}
\alias{browse_show_collection}
\title{Show collection of MSigDB database}
\usage{
browse_show_collection()
}
\value{
show all collection in MSigDB in web page http://www.gsea-msigdb.org/gsea/msigdb/genesets.jsp.
For chromosome, we should treat as collection together.
}
\description{
Show collection of MSigDB database
}
\examples{
\donttest{
browse_show_collection()
}

}
