% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_query.R
\name{local_msig}
\alias{local_msig}
\title{Query gene set names from local data}
\usage{
local_msig(
  geneset = NULL,
  description = NULL,
  collection = NULL,
  sub_collection = NULL,
  organism = NULL,
  contributor = NULL,
  contributor_org = NULL,
  author = NULL,
  chip = NULL,
  gene = NULL,
  geneset_fuzzy = NULL,
  collection_fuzzy = NULL,
  sub_collection_fuzzy = NULL,
  organism_fuzzy = NULL,
  contributor_fuzzy = NULL,
  contributor_org_fuzzy = NULL,
  author_fuzzy = NULL,
  chip_fuzzy = NULL,
  gene_fuzzy = NULL
)
}
\arguments{
\item{geneset}{one sql format character for exact match}

\item{description}{one sql format character for exact match}

\item{collection}{one sql format character for exact match}

\item{sub_collection}{one sql format character for exact match}

\item{organism}{one sql format character for exact match}

\item{contributor}{one sql format character for exact match}

\item{contributor_org}{one sql format character for exact match}

\item{author}{one sql format character for exact match}

\item{chip}{one sql format character for exact match}

\item{gene}{one sql format character for exact match}

\item{geneset_fuzzy}{fuzzy match}

\item{collection_fuzzy}{fuzzy match}

\item{sub_collection_fuzzy}{fuzzy match}

\item{organism_fuzzy}{fuzzy match}

\item{contributor_fuzzy}{fuzzy match}

\item{contributor_org_fuzzy}{fuzzy match}

\item{author_fuzzy}{fuzzy match}

\item{chip_fuzzy}{fuzzy match}

\item{gene_fuzzy}{fuzzy match}
}
\value{
one dataframe with attribute of msig_local.
}
\description{
Query gene set names from local data
}
\examples{
x <- local_msig('IMMUNE_RESPONSE')
x <- local_msig('IMMUNE_RESPONSE|IMMUNE_SYSTEM_PROCESS')
}
