% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_gst_detail_detail.R
\name{msig_detail}
\alias{msig_detail}
\title{Retrieve detail information of gene set}
\usage{
msig_detail(...)
}
\arguments{
\item{...}{one or more gene set names, which can be little or capital.}
}
\value{
Print detail information about the geneset, number of genes and return all gene names.
}
\description{
Retrieve detail information of gene set
}
\examples{
\donttest{
d <- msig_detail('izadpanah_stem_cell_adipose_vs_bone_dn',
                 'AAACCAC_MIR140')
}

}
