% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_download.R
\name{msig_download}
\alias{msig_download}
\title{Download MsigDB database}
\usage{
msig_download(version)
}
\arguments{
\item{version}{version}
}
\value{
dowanload the data to local PC
}
\description{
Download MsigDB database
}
