% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msig_gene.R, R/msig_gene.character.R,
%   R/msig_gene.data.frame.R
\name{msig_gene}
\alias{msig_gene}
\alias{msig_gene.character}
\alias{msig_gene.data.frame}
\title{Retrieve gene by Gene set Name}
\usage{
msig_gene(..., list = TRUE, info = TRUE)

\method{msig_gene}{character}(..., list = TRUE, info = TRUE)

\method{msig_gene}{data.frame}(..., list = TRUE, info = TRUE)
}
\arguments{
\item{...}{one or more geneset names, which can be little or capital.}

\item{list}{logical, default is FALSE, whether to show result by list.}

\item{info}{logical, whether to show information about gene set.}
}
\value{
Print detail information about the geneset, number of genes and return all gene names.
}
\description{
Retrieve gene by Gene set Name
}
\examples{
\donttest{
genes <- msig_gene('izadpanah_stem_cell_adipose_vs_bone_dn',
                   'REACTOME_DEGRADATION_OF_AXIN')
genes |>
    msig_view()
}


}
