% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_update.R
\name{msig_update}
\alias{msig_update}
\title{Update local MsigDB database}
\usage{
msig_update(xml = NULL, version = NULL)
}
\arguments{
\item{xml}{msigdb xml file}

\item{version}{version, if missing, the latest version will be used}
}
\value{
update local MsigDB database
}
\description{
Update local MsigDB database
}
