% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_view.R
\name{msig_view}
\alias{msig_view}
\title{View data in viewer panel}
\usage{
msig_view(x, ...)
}
\arguments{
\item{x}{dataframe}

\item{...}{one or more hightliht words}
}
\value{
open data in view panel in rstudio
}
\description{
View data in viewer panel
}
\examples{
#' browse_msig('immune') |>
    msig_view('response')
}
