% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_NewMsigDB.R
\name{NewMsigDB}
\alias{NewMsigDB}
\title{Create NewMsigDB object for new versions of MsigDB database}
\usage{
NewMsigDB(xml)
}
\arguments{
\item{xml}{path of xml msigdb file path}
}
\value{
dataframe which can be used inner package
}
\description{
Create NewMsigDB object for new versions of MsigDB database
}
