% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msigdb_xml.R
\name{read_msigdb_xml}
\alias{read_msigdb_xml}
\title{read MSigDB xml data}
\usage{
read_msigdb_xml(xml)
}
\arguments{
\item{xml}{xml data path}
}
\value{
one dataframe contains gene infomation
}
\description{
read MSigDB xml data
}
