% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_gst_detail_related.R
\name{related_geneset}
\alias{related_geneset}
\title{Query related gene sets}
\usage{
related_geneset(geneSetName)
}
\arguments{
\item{geneSetName}{one gene set name}
}
\value{
related gene sets from gene set detailed information table
}
\description{
Query related gene sets
}
\examples{
\donttest{
x <- related_geneset('AAANWWTGC_UNKNOWN')
x |>
    msig_filt('unknown') |>
    msig_view('ttt')
}

}
