% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_gst_sch_2_query.R
\name{search_msig}
\alias{search_msig}
\title{Query MSigDB database by cookie}
\usage{
search_msig(keywords, collection = "", organism = "", contributor = "", email)
}
\arguments{
\item{keywords}{one keywords see Detail field}

\item{collection}{one or more collections}

\item{organism}{one or more organisms}

\item{contributor}{one or more contributors}

\item{email}{email that registered for MSigDB database.}
}
\value{
dataframe contains name, description and so on.
}
\description{
Query MSigDB database by cookie
}
\examples{
\donttest{
email <- 'your email'
x <- search_msig('immune & response')
x |>
    msig_filt('system') |>
    msig_view('C2')
}

}
