% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_gst_sch_1_show_collection.R
\name{search_show_collection}
\alias{search_show_collection}
\title{Show collctions for msigdb_search()}
\usage{
search_show_collection(email)
}
\arguments{
\item{email}{email that registered for MSigDB database.}
}
\value{
collections from MsigDB website.
}
\description{
Show collctions for msigdb_search()
}
\examples{
\donttest{
search_show_collection("your email")
# or
email <- 'your email'
search_show_collection()
}
}
