% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_gst_sch_1_show_contributor.R
\name{search_show_contributor}
\alias{search_show_contributor}
\title{Show contributor for msigdb_search()}
\usage{
search_show_contributor(email)
}
\arguments{
\item{email}{email that registered for MSigDB database.}
}
\value{
contributors from MsigDB website.
}
\description{
Show contributor for msigdb_search()
}
\examples{
\donttest{
search_show_contributor("your email")
# or
email <- 'your email'
search_show_contributor()
}
}
