% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_gst_sch_1_show_organism.R
\name{search_show_organism}
\alias{search_show_organism}
\title{Show organism for msigdb_search()}
\usage{
search_show_organism(email)
}
\arguments{
\item{email}{email that registered for MSigDB database.}
}
\value{
organisms from MsigDB website.
}
\description{
Show organism for msigdb_search()
}
\examples{
\donttest{
search_show_organism("your email")
# or
email <- 'your email'
search_show_organism()
}
}
