% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_show_collection.R
\name{show_local_collection}
\alias{show_local_collection}
\title{Show collections of local MsigDB database}
\usage{
show_local_collection()
}
\value{
A dataframe contains 2 columns. The first column is the name of the collection.
The second column is the number of frequencies it has.
}
\description{
Show collections of local MsigDB database
}
\examples{
show_local_collection()

}
