% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_local_contributor.R
\name{show_local_contributor}
\alias{show_local_contributor}
\title{Show contributors of local MsigDB database}
\usage{
show_local_contributor()
}
\value{
A dataframe contains 2 columns. The first column is the name of the contributor.
The second column is the number of frequencies it has.
}
\description{
Show contributors of local MsigDB database
}
\examples{
show_local_contributor()

}
