% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_local_contributor_org.R
\name{show_local_contributor_org}
\alias{show_local_contributor_org}
\title{Show contributor_orgs of local MsigDB database}
\usage{
show_local_contributor_org()
}
\value{
A dataframe contains 2 columns. The first column is the name of the contributor_org.
The second column is the number of frequencies it has.
}
\description{
Show contributor_orgs of local MsigDB database
}
\examples{
show_local_contributor_org()

}
