% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_show_data.R
\name{show_local_data}
\alias{show_local_data}
\title{Show local data used in this package}
\usage{
show_local_data()
}
\value{
data used inner this package
}
\description{
Show local data used in this package
}
\examples{
show_local_data()

}
