% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_local_organism.R,
%   R/show_local_sub_collection.R
\name{show_local_sub_collection}
\alias{show_local_sub_collection}
\title{Show sub_collections of local MsigDB database}
\usage{
show_local_sub_collection()

show_local_sub_collection()
}
\value{
A dataframe contains 2 columns. The first column is the name of the sub_collection.
The second column is the number of frequencies it has.

A dataframe contains 2 columns. The first column is the name of the sub_collection.
The second column is the number of frequencies it has.
}
\description{
Show sub_collections of local MsigDB database

Show sub_collections of local MsigDB database
}
\examples{
show_local_sub_collection()

show_local_sub_collection()

}
