% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_gst_detail_similarity.R
\name{similarity_geneset}
\alias{similarity_geneset}
\title{Query similarity gene sets}
\usage{
similarity_geneset(geneSetName)
}
\arguments{
\item{geneSetName}{one gene set name}
}
\value{
similarity gene sets
}
\description{
Query similarity gene sets
}
\examples{
\donttest{
x <- similarity_geneset('REACTOME_DEGRADATION_OF_AXIN')
x |>
    msig_view()

}
}
